Fix building without zopfli
